+++ startup/GafferScene/arnoldAttributes.py
@@ -71,9 +71,9 @@ Gaffer.Metadata.registerValues( {
 		"description",
 		"""
 		The lights that cause this object to cast shadows.
-		Accepts a set expression or a space separated list of
-		lights. Use \"defaultLights\" to refer to all lights that
-		contribute to illumination by default.
+
+		> Caution : This attribute has been superceded and will be removed. Use
+		> the standard `shadowedLights` attribute instead.
 		""",
 		"label", "Shadow Group",
 		"ui:scene:acceptsSetExpression", True,